refactor: validaor함수 분리
