+++ hooks/src/lib/index.ts
@@ -1,4 +1,4 @@
-export { default as useCardNumbers } from "../hooks/useCardNumbers";
+export { default as useCardNumbers } from "../hooks/useCardNumbers/useCardNumbers";
 export { default as useCardCompany } from "../hooks/useCardCompany";
 export { default as useExpiryPeriod } from "../hooks/useExpiryPeriod";
 export { default as useRestrictedState } from "../hooks/useRestrictedState";
